+++ packages/backend-data/src/factory.ts
@@ -1,5 +1,6 @@
 import { IConstruct } from 'constructs';
 import {
+  AmplifyFunction,
   AuthResources,
   BackendOutputStorageStrategy,
   ConstructContainerEntryGenerator,
@@ -109,9 +110,11 @@ class DataGenerator implements ConstructContainerEntryGenerator {
     let amplifyGraphqlDefinition;
     let jsFunctions: JsResolver[] = [];
     let functionSchemaAccess: FunctionSchemaAccess[] = [];
+    let lambdaFunctions: Record<string, ConstructFactory<AmplifyFunction>> = {};
     try {
       if (isModelSchema(this.props.schema)) {
-        ({ jsFunctions, functionSchemaAccess } = this.props.schema.transform());
+        ({ jsFunctions, functionSchemaAccess, lambdaFunctions } =
+          this.props.schema.transform());
       }
       amplifyGraphqlDefinition = convertSchemaToCDK(this.props.schema);
     } catch (error) {
@@ -121,7 +124,9 @@ class DataGenerator implements ConstructContainerEntryGenerator {
           message:
             error instanceof Error
               ? error.message
-              : 'Cannot covert user schema',
+              : 'Failed to parse schema definition.',
+          resolution:
+            'Check your data schema definition for syntax and type errors.',
         },
         error instanceof Error ? error : undefined
       );
@@ -152,7 +157,8 @@ class DataGenerator implements ConstructContainerEntryGenerator {
           message:
             error instanceof Error
               ? error.message
-              : 'Cannot covert authorization modes',
+              : 'Failed to parse authorization modes.',
+          resolution: 'Ensure the auth rules on your schema are valid.',
         },
         error instanceof Error ? error : undefined
       );
@@ -171,6 +177,7 @@ class DataGenerator implements ConstructContainerEntryGenerator {
             error instanceof Error
               ? error.message
               : 'Failed to validate authorization modes',
+          resolution: 'Ensure the auth rules on your schema are valid.',
         },
         error instanceof Error ? error : undefined
       );
@@ -181,10 +188,12 @@ class DataGenerator implements ConstructContainerEntryGenerator {
       this.props.authorizationModes
     );
 
-    const functionNameMap = convertFunctionNameMapToCDK(
-      this.getInstanceProps,
-      this.props.functions ?? {}
-    );
+    const propsFunctions = this.props.functions ?? {};
+
+    const functionNameMap = convertFunctionNameMapToCDK(this.getInstanceProps, {
+      ...propsFunctions,
+      ...lambdaFunctions,
+    });
     const amplifyApi = new AmplifyData(scope, this.defaultName, {
       apiName: this.props.name,
       definition: amplifyGraphqlDefinition,