+++ packages/backend-storage/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/backend-storage
 
+## 0.6.0-beta.4
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.4.0-beta.2
+
 ## 0.6.0-beta.3
 
 ### Minor Changes